requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
aiosmtplib==4.0.1
annotated-types==0.7.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
frozenlist==1.7.0
idna==3.10
importlib_metadata==8.7.0
inflect==7.5.0
magic-filter==1.0.12
more-itertools==10.7.0
multidict==6.4.4
packaging==25.0
pillow==11.2.1
propcache==0.3.2
pydantic==2.11.5
pydantic_core==2.33.2
pytelebot==0.0.1
pyTelegramBotAPI==4.13.0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.4
telebotapi==0.9.9.3
typeguard==4.4.4
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
yarl==1.20.0
zipp==3.23.0
